+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+### Fixed
+
+- Enable minimize in `webpack.config.js`
+
 ## [2.4.2] - 2022-12-21
 
 ### Changed